+++ mafia-app-api/src/main/java/com/mafia/api/repository/MafiaClubRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mafia.api.models.MafiaClub;
 
-public interface MafiaClubRepository extends JpaRepository<MafiaClub, Long>{
+public interface MafiaClubRepository extends JpaRepository<MafiaClub, Integer>{
     
 }
